Re-organize files into pages and backend folders
