+++ lib/pages/home_page.dart
@@ -1,10 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:mup_app/login.dart';
+import 'package:mup_app/pages/login.dart';
 import 'package:mup_app/pages/frequency_profile.dart';
 import 'package:mup_app/pages/my_account.dart';
 import 'package:mup_app/pages/add_new_device.dart';
 import 'package:mup_app/pages/system_health.dart';
-
 
 class MyApp extends StatelessWidget {
   // This widget is the root of your application.
